Moved normalizeDomainInput into the domain utils
