+++ dashboard/src/app/(app)/(protected)/dashboards/CreateDashboardDialog.tsx
@@ -16,7 +16,8 @@ import { Label } from '@/components/ui/label';
 import { Tooltip, TooltipTrigger, TooltipContent, TooltipProvider } from '@/components/ui/tooltip';
 import { Plus, Lock } from 'lucide-react';
 import { createDashboardAction, getUserDashboardStatsAction } from '@/app/actions/dashboard/dashboard.action';
-import { domainValidation, normalizeDomainInput } from '@/entities/dashboard/dashboard.entities';
+import { domainValidation } from '@/entities/dashboard/dashboard.entities';
+import { normalizeDomainInput } from '@/utils/domainValidation';
 import { useBARouter } from '@/hooks/use-ba-router';
 import { useTranslations } from 'next-intl';
