+++ dashboard/src/app/(app)/[locale]/(onboarding)/onboarding/steps/WebsiteSetup.tsx
@@ -4,7 +4,8 @@ import { useState, useTransition, useCallback, useMemo, Dispatch } from 'react';
 import { Button } from '@/components/ui/button';
 import { Label } from '@/components/ui/label';
 import { completeOnboardingAndCreateDashboardAction } from '@/app/actions/account/onboarding.action';
-import { domainValidation, normalizeDomainInput } from '@/entities/dashboard/dashboard.entities';
+import { domainValidation } from '@/entities/dashboard/dashboard.entities';
+import { normalizeDomainInput } from '@/utils/domainValidation';
 import { toast } from 'sonner';
 import { PrefixInput } from '@/components/inputs/PrefixInput';
 import { useTranslations, useLocale } from 'next-intl';
